requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-import-export==3.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0